experiment/requirements.txt
pyrtlsdr==0.3.0
numpy==1.24.0
scipy==1.10.0
matplotlib==3.7.0
adafruit-circuitpython-rfm9x==2.2.0
